Reorganization of module.
